Wyapoint 4 Split out more stuff into seperate files to reduce glut of formmap
Todo: Zoom panning, messages on movement
