+++ EDDiscovery/3DMap/FormMap.resx
@@ -741,7 +741,7 @@
   <data name="buttonCenter.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7O
-        HOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAAAsSAAALEgHS3X78AAACzklEQVRIS61WTYwMQRTu2SV+
+        HOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAAAsRAAALEQF/ZF+RAAACzklEQVRIS61WTYwMQRTu2SV+
         4581s12vqqurd1gikhEHCePGwVEkjpIlmyAcHESEuLkJQUQcCXHwk/g7SFxICBIX4YSInwiXJfEzsvhe
         zdvo7ek2Y7NfUpn+Xtd7r+p7r6on6BT98/unO22+xmS2iWl80Vfum4cEv53We8U0PtBaL4yVWpJOQESz
         E62X43WpOWsMiIlWOaJbCPoLv1fSCRxFG71cyrx0yuyq1WoTxa096kF9AnQ+YhU1EOR4HMaO7VmJyuXy
